Prevent type from being set in Storybook urls in Cypress tests
The normal flow of rendering materials is that if no material type is
set then one will be detected automatically and the url with be updated
accordingly.

If the type in the url does not match a for the current the current 
material e.g. then rendering of the material app can stop at the 
skeleton screen. This can be caused by updated material types for the
material or updated material names.

To avoid this we configure Eslint to warn if such urls are called
through AST analysis and regular expressions.