Implement alternate hook for retrieving reservations
FBS will return normal and parallel reservations in the same data
structure with one entry per reserved record/Faust id. We want to treat
parallel reservations of multiple records as one reservation as it will
be perceived as such by the end user.

Implement unit test to show that things are working as expected.